rename and reorganize message_handler to output subfolder and add config file loading
